only pass the --skip argument to tests if a dedicated test driver is specified
since ewoms-eclsimulator runs its tests on bare metal all its
conditional tests where the condition was not met failed before this
instead of being skipped.

Note that this fix is a bit hacky; the assumption is that the test
driver is always there and under full control of the user, i.e., it
can always be made to accept the --skip option.