fix the cmake test for std::filesystem
the sh** hit the fan when trying to compile ewoms-eclio with GCC 8.3:
since the eclio library uses stuff which requires linking against the
libstdc++fs library, but the cmake test did not check any of this
functionality, it thought that no special library was needed.  This
resulted in a linker error during the build.