build system: use CONFIGURE_DEPENDS flag for glob expressions on CMake >= 3.12
on supported target systems (i.e., make), this causes the glob
expressions to be evaluated at runtime. while this is not a 100%
bullet-proof solution, it gets rid of having to manually run cmake
after an update in about 95% of all cases without having to maintain
burdensome global lists of these files. (OPM chose to do it this way
and their lists have a few errors like missing files, etc...)